+++ lib/maru/builder/exceptions.ex
@@ -3,64 +3,71 @@ defmodule Maru.Builder.Exceptions do
   Handle exceptions of current router.
   """
 
-  @doc false
-  defmacro rescue_from(:all, [as: error_var], [do: block]) do
-    quote do
-      @exceptions {:all, unquote(Macro.escape error_var), unquote(Macro.escape block)}
-    end
-  end
-
-  defmacro rescue_from(error, [as: error_var], [do: block]) do
-    quote do
-      @exceptions {unquote(error), unquote(Macro.escape error_var), unquote(Macro.escape block)}
-    end
-  end
+  alias Maru.Struct.Exception
 
   @doc false
-  defmacro rescue_from(:all, [do: block]) do
+  defmacro rescue_from(errors, [as: error_var], [do: block]) do
+    errors = format_errors(errors)
     quote do
-      @exceptions {:all, unquote(Macro.escape block)}
-    end
-  end
-
-  defmacro rescue_from(error, [do: block]) do
-    quote do
-      @exceptions {unquote(error), unquote(Macro.escape block)}
-    end
-  end
-
-
-  @doc false
-  def make_rescue_block({:all, block}) do
-    quote do
-      _ ->
-        var!(conn) = Maru.Helpers.Response.get_maru_conn
-        unquote(block)
-    end
-  end
-
-  def make_rescue_block({error, block}) do
-    quote do
-      unquote(error) ->
-        var!(conn) = Maru.Helpers.Response.get_maru_conn
-        unquote(block)
+      @exceptions %Exception{
+        errors:    unquote(errors),
+        error_var: unquote(Macro.escape error_var),
+        block:     unquote(Macro.escape block),
+      }
     end
   end
 
   @doc false
-  def make_rescue_block({:all, error_var, block}) do
+  defmacro rescue_from(errors, [do: block]) do
+    errors = format_errors(errors)
     quote do
-      unquote(error_var) ->
-        var!(conn) = Maru.Helpers.Response.get_maru_conn
-        unquote(block)
+      @exceptions %Exception{
+        errors: unquote(errors),
+        block:  unquote(Macro.escape block),
+      }
     end
   end
 
-  def make_rescue_block({error, error_var, block}) do
+  defmacro rescue_from(errors, [with: function]) do
+    errors = format_errors(errors)
     quote do
-      unquote(error_var) in unquote(error) ->
-        var!(conn) = Maru.Helpers.Response.get_maru_conn
-        unquote(block)
+      @exceptions %Exception{
+        errors:    unquote(errors),
+        error_var: Macro.var(:maru_exception, nil),
+        function:  unquote(function),
+      }
+    end
+  end
+
+  defp format_errors(:all),  do: :all
+  defp format_errors(errors)
+  when is_list(errors),      do: errors
+  defp format_errors(error), do: [error]
+
+  @doc false
+  def make_rescue_block(%Exception{}=exception) do
+    var   = make_variable(exception.errors, exception.error_var)
+    block = make_block(exception.function, exception.error_var, exception.block)
+    quote do
+      unquote(var) -> unquote(block)
+    end
+  end
+
+  defp make_variable(:all,   nil), do: (quote do _ end)
+  defp make_variable(:all,   var), do: var
+  defp make_variable(errors, nil), do: errors
+  defp make_variable(errors, var), do: (quote do unquote(var) in unquote(errors) end)
+
+  defp make_block(nil, _, block) do
+    quote do
+      var!(conn) = Maru.Helpers.Response.get_maru_conn
+      unquote(block)
+    end
+  end
+
+  defp make_block(function, var, nil) do
+    quote do
+      unquote(function)(Maru.Helpers.Response.get_maru_conn, unquote(var))
     end
   end