+++ lib/maru/struct/exception.ex
@@ -1,0 +1,9 @@
+defmodule Maru.Struct.Exception do
+  @moduledoc false
+
+  defstruct errors:    nil,
+            error_var: nil,
+            function:  nil,
+            block:     nil
+
+end