+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   * add :keep_blank option for params
   * add :given DSL for dependent parameters
   * make `rescue_from` works for all maru router
+  * support `with` option for `rescue_from` DSL
+
 * Bugfix
   * allow set optional params to false and blank value by default